Move out vg deconstruct
