+++ Makefile
@@ -166,6 +166,7 @@ SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/stats_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/vectorize_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/filter_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/version_main.o
+SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/deconstruct_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/locify_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/paths_main.o
 SUBCOMMAND_OBJ += $(SUBCOMMAND_OBJ_DIR)/join_main.o
@@ -630,7 +631,9 @@ $(SUBCOMMAND_OBJ_DIR)/align_main.o: $(SUBCOMMAND_SRC_DIR)/align_main.cpp $(SUBCO
 
 $(SUBCOMMAND_OBJ_DIR)/surject_main.o: $(SUBCOMMAND_SRC_DIR)/surject_main.cpp $(SUBCOMMAND_SRC_DIR)/subcommand.hpp $(SRC_DIR)/vg.hpp $(SRC_DIR)/stream.hpp $(SRC_DIR)/mapper.hpp $(DEPS)
 
-$(SUBCOMMAND_OBJ_DIR)/version_main.o: $(SUBCOMMAND_SRC_DIR)/version_main.cpp $(SUBCOMMAND_SRC_DIR)/subcommand.hpp $(SRC_DIR)/version.hpp $(DEPS)
+$(SUBCOMMAND_OBJ_DIR)/version_main.o: $(SUBCOMMAND_SRC_DIR)/version_main.cpp $(SUBCOMMAND_SRC_DIR)/subcommand.hpp $(SRC_DIR)/version.hpp $(SRC_DIR)/deconstructor.hpp $(DEPS)
+
+$(SUBCOMMAND_OBJ_DIR)/deconstruct_main.o: $(SUBCOMMAND_SRC_DIR)/deconstruct_main.cpp $(SUBCOMMAND_SRC_DIR)/subcommand.hpp $(SRC_DIR)/vg.hpp $(DEPS)
 
 $(SUBCOMMAND_OBJ_DIR)/locify_main.o: $(SUBCOMMAND_SRC_DIR)/locify_main.cpp $(SUBCOMMAND_SRC_DIR)/subcommand.hpp $(SRC_DIR)/vg.hpp $(SRC_DIR)/index.hpp $(SRC_DIR)/convert.hpp $(DEPS)
